more CS50 projects
These are more projects from the cs50 class.  The class originally started writing code in C and then moved on to SQL, python, and eventually html and JavaScript.  There are a few projects like readability, Mario, and cash that were written first in C and then written again in python to get the hang of the new language.